% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_taxonomy.R
\name{bin_taxonomy}
\alias{bin_taxonomy}
\title{Bin ASV count table at specified level of taxonomic classification.}
\usage{
bin_taxonomy(asv.table, meta.data, tax.level)
}
\arguments{
\item{asv.table}{ASV count table named by Silva-138 L7 taxonomies. Ideally rarefied and filtered as necessary.
First column must be named "\code{sample-id}" and must contain unique identifiers.}

\item{meta.data}{Sample metadata corresponding to asv.table.}

\item{tax.level}{Level of taxonomic classification at which to calculate relative abundance.}
}
\value{

}
\description{
Bin ASV count table at specified level of taxonomic classification.
}
\examples{
bin_taxonomy(asv_table, metadata, "class")
}
