% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_taxa_barplot.R
\name{build_taxa.barplot}
\alias{build_taxa.barplot}
\title{Build summary taxa barplots from network abundance data}
\usage{
build_taxa.barplot(
  plotdata,
  tax.level,
  dendrogram,
  fill.by = "phylum",
  interactive = TRUE,
  out.dr,
  metadata.cols,
  other.variable
)
}
\arguments{
\item{plotdata}{Output from \code{\link{ebc_compositions}}}

\item{tax.level}{Taxonomic level}

\item{dendrogram}{Output from \code{\link{build_dendrograms}}}

\item{fill.by}{Fill option for summary barplot. Can be one of \code{"phylum"} (to fill by phylum abundance), \code{"ebc"} (to fill by edge between cluster abundance), or \code{"grouping"} (to fill by family-level functional groupings). \code{"grouping"} is only available for family-level and below.}

\item{interactive}{If \code{TRUE}, plot exported as interactive HTML file. If \code{FALSE}, plot exported as PDF. Default = \code{TRUE}.}

\item{out.dr}{Output directory}

\item{other.variable}{}
}
\value{

}
\description{
Build summary taxa barplots from network abundance data
}
