% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ebc_abun.R
\name{calc_ebc_abun}
\alias{calc_ebc_abun}
\title{Calculate abundance of edge between clusters}
\usage{
calc_ebc_abun(ebc.nodes, binned.taxonomy, metadata.cols)
}
\arguments{
\item{metadata.cols}{}
}
\value{
\strong{deprecated}
}
\description{
Calculate abundance of edge between clusters
}
