% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ebc_comps.R
\name{calc_ebc_comps}
\alias{calc_ebc_comps}
\title{Calculate compositions of edge between clusters based on defined metadata column(s)}
\usage{
calc_ebc_comps(ebc.nodes, binned.taxonomy, tax.level, metadata.cols)
}
\value{
\strong{Deprecated!} This function will be removed in a future version of \code{bngal}. Please use \code{\link{ebc_compositions}} instead.
}
\description{
Calculate compositions of edge between clusters based on defined metadata column(s)
}
