% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cores.R
\name{check_cores}
\alias{check_cores}
\title{Check number of available CPUs}
\usage{
check_cores(num.cores = NULL)
}
\arguments{
\item{num.cores}{User-provided number of cores to use. If the variable is left as NULL (or missing entirely), bngal will automatically detect all available cores on the (non-Windows) machine. If a Windows platform is detected, num.cores is set to 1.}
}
\description{
This function checks for the number of available CPUs (cores) available
on the user's machine. This is formatted for parallel processing on a
SLURM-directed HPC system, but any *nix-like machine can multithread here
as well. If a Windows OS is detected, this function automatically returns \code{1}
to run on a single core.
}
\details{
It is recommended to save this value to be called from multithreaded bngal
functions such as \code{bngal::color_nodes}
}
\examples{
n.cores <- bngal::check_cores()
}
