% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_nodes.R
\name{color_nodes}
\alias{color_nodes}
\title{Color network nodes by edge between cluster ID and phylum}
\usage{
color_nodes(clusters.to.color, phylum.colors)
}
\arguments{
\item{clusters.to.color}{Output from \code{\link[=get_ebc_clusters]{get_ebc_clusters()}}}

\item{phylum.colors}{\emph{Optional} Dataframe with hex color codes assigned to
each phylum in the Silva 16SrRNA gene database (v. 138). Default color scheme
provided, but if a custom color scheme is required, columns must be the
following:
\itemize{
\item \code{phylum}: name of phylum in Silva database
\item \code{hex.color}: color hexcode, including \verb{#}
\item \code{order}: desired order for downstream plots
}}
}
\value{

}
\description{
Color network nodes by edge between cluster ID and phylum
}
