% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Compute correlation matrix}
\usage{
corr_matrix(filtered.matrix, correlation, cores)
}
\arguments{
\item{filtered.matrix}{Output from \verb{[bngal::prepare_corr_data]}}

\item{correlation}{See \verb{[Hmisc::rcorr()]}}
}
\description{
Calculate correlation matrix from filtered data. This is essentially a
convenience wrapper around \verb{[Hmisc::rcorr()]}.
}
