% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebc_compositions.R
\name{ebc_compositions}
\alias{ebc_compositions}
\title{Calculate compositions of edge between clusters based on defined metadata column(s)}
\usage{
ebc_compositions(
  ebc.nodes,
  binned.taxonomy,
  alpha.div,
  tax.level,
  metadata,
  metadata.cols,
  sub.comms
)
}
\arguments{
\item{ebc.nodes}{Output from \code{\link[=extract_node_data]{extract_node_data()}}}

\item{binned.taxonomy}{Output from \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{tax.level}{Taxonomic level of classification at which to calculate EBC compositions}

\item{sub.comms}{\emph{(Optional)} Metadata column by which to split data into subcommunities}
}
\description{
Calculate compositions of edge between clusters based on defined metadata column(s)
}
