% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cluster_data.R
\name{export_cluster_data}
\alias{export_cluster_data}
\title{Export filtered edge between cluster abundance data}
\usage{
export_cluster_data(ebc.comp.data, quantile.cutoff)
}
\arguments{
\item{ebc.comp.data}{Output from \code{bngal::calc_ebc_comps()}}

\item{quantile.cutoff}{(numeric) Quantile distribution cutoff for coreness/abundance data:
\itemize{
\item 1: 0-25\%
\item 2: 25-50\%
\item 3: 50-75\%
\item 4: 75-100\%
}}
}
\description{
This function returns a list of nodes (i.e., taxa) at each level of
taxonomic classification, split by their statistical distributions into
four categories listed below (defined by the option \code{quantile.cutoff}) :
\itemize{
\item \code{high.core_high.abun} : Nodes with high coreness and high abundance
\item \code{high.core_low.abun}  : Nodes with high coreness and low abundance
\item \code{low.core_high.abun}  : Nodes with low coreness and high abundance
\item \code{low.core_low.abun}   : Nodes with low coreness and low abundance
}
}
\examples{
export_cluster_data(ebc_composition_data, 4)
}
