% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ebc_taxa_summary.R
\name{export_ebc_taxa_summary}
\alias{export_ebc_taxa_summary}
\title{Export EBC and taxonomic abundance summary data to CSV file}
\usage{
export_ebc_taxa_summary(binned.taxonomy, ebc.nodes.abun, tax.level, out.dr)
}
\arguments{
\item{binned.taxonomy}{Output from \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{ebc.nodes.abun}{Output from \code{\link[=ebc_compositions]{ebc_compositions()}}}

\item{tax.level}{Taxonomic level at which to summarize data.}

\item{out.dr}{Output directory. Results will be saved to the \code{"network-summary-tables"} subfolder.}
}
\value{

}
\description{
Export EBC and taxonomic abundance summary data to CSV file
}
