% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_network_data.R
\name{export_network_data}
\alias{export_network_data}
\title{Export network node and ebc cluster data}
\usage{
export_network_data(node.color.data, tax.level, subset.column, out.dr)
}
\arguments{
\item{node.color.data}{Output from \code{\link[=color_nodes]{color_nodes()}}}

\item{tax.level}{See \code{\link{prepare_network_data}}}

\item{out.dr}{Data output directory (does not have to exists)}
}
\value{

}
\description{
Export network node and ebc cluster data
}
