% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_node_data.R
\name{extract_node_data}
\alias{extract_node_data}
\title{Import network data}
\usage{
extract_node_data(network.data)
}
\arguments{
\item{network.data}{Output from Output from \code{\link[=export_network_data]{export_network_data()}}}
}
\value{

}
\description{
Import network data
}
