% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edges.R
\name{generate_edges}
\alias{generate_edges}
\title{Generate network edges}
\usage{
generate_edges(corr.matrix, correlation, node.ids)
}
\arguments{
\item{corr.matrix}{Output from \code{\link[=prepare_corr_matrix]{prepare_corr_matrix()}}}

\item{correlation}{Metric for pairwise comparisions (see \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}})}

\item{node.ids}{Output from \code{\link[=get_node_ids]{get_node_ids()}}}
}
\value{

}
\description{
Generate network edges
}
\examples{
generate_edges(corr_matrix, "spearman", nodes)
}
