% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebc_clusters.R
\name{get_ebc_clusters}
\alias{get_ebc_clusters}
\title{Return taxonomic compositions of each edge between cluster}
\usage{
get_ebc_clusters(prepro.data, ebc.member.ids, igraph.obj, core.ness = "all")
}
\arguments{
\item{prepro.data}{Output from \code{\link[=prepro_net_features]{prepro_net_features()}}}

\item{ebc.member.ids}{Output from \code{\link[=get_ebc_member_ids]{get_ebc_member_ids()}}}

\item{igraph.obj}{Output from \code{\link[=get_igraph]{get_igraph()}}}

\item{core.ness}{See \code{mode} argument from \code{\link[igraph:coreness]{igraph::coreness()}}. Default = \code{"all"}}
}
\value{

}
\description{
Return taxonomic compositions of each edge between cluster
}
\examples{
get_ebc_clusters()
}
