% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebc_member_ids.R
\name{get_ebc_member_ids}
\alias{get_ebc_member_ids}
\title{Get edge between cluster IDs}
\usage{
get_ebc_member_ids(betweenness.)
}
\arguments{
\item{betweenness.}{Output from \code{\link[=get_edge_betweenness]{get_edge_betweenness()}}}
}
\value{

}
\description{
Wrapper function around \code{\link[igraph:communities]{igraph::membership()}}
}
\examples{
ebcs <- bngal::get_edge_betweenness(igraph_obj)
bngal::get_ebc_member_ids(ebcs)
}
