% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_igraph.R
\name{get_igraph}
\alias{get_igraph}
\title{Create igraph object}
\usage{
get_igraph(prepro.data, directed. = FALSE)
}
\arguments{
\item{prepro.data}{Output from \code{\link{prepro_net_features}}}

\item{directed.}{See \code{directed} option from \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame}}}
}
\value{

}
\description{
This is function is a wrapper around \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame()}} to
generate igraph objects for each subcommunity defined earlier in the
\code{bngal} pipeline.
}
\examples{
get_igraph(prepped_nodes_edges)
}
