% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_ids.R
\name{get_node_ids}
\alias{get_node_ids}
\title{Get network node IDs}
\usage{
get_node_ids(prepared.data, corr.matrix)
}
\arguments{
\item{prepared.data}{Output from \code{\link{prepare_network_data}}}

\item{corr.matrix}{Output from \code{\link[=prepare_corr_data]{prepare_corr_data()}}}
}
\description{
This function yields a dataframe containing each unique, quality-filtered
node to be used in constructing a correlation network.
}
