% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_network_data.R
\name{load_network_data}
\alias{load_network_data}
\title{Load exported network data}
\usage{
load_network_data(network.dir)
}
\arguments{
\item{network.dir}{Full path to network data output from \verb{[bngal::export_network_data]}}
}
\value{

}
\description{
Load exported network data
}
