% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Make matrix from long-form microbiome data}
\usage{
make_matrix(long_data, count_column_name, row_names)
}
\arguments{
\item{long_data}{Output from \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{count_column_name}{name of count column}

\item{row_names}{column by which to name table rows (default = \code{"sample-id"})}
}
\description{
Make matrix from long-form microbiome data
}
\examples{
make_matrix(long_data, "rel_abun", )
}
