% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_comp.R
\name{plot_core_comp}
\alias{plot_core_comp}
\title{Plot edge between cluster compositions against node coreness}
\usage{
plot_core_comp(ebc.nodes.abun, tax.level, metadata, fill.by)
}
\arguments{
\item{fill.by}{}
}
\value{

}
\description{
Plot edge between cluster compositions against node coreness
}
