% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_corr_data.R
\name{prepare_corr_data}
\alias{prepare_corr_data}
\title{Filter preprocessed data for correlation matrix}
\usage{
prepare_corr_data(prepared.data, obs.cutoff, transformation, out.dr, cores = 1)
}
\arguments{
\item{prepared.data}{Output from \code{\link{prepare_network_data}} or \code{\link{split_network_data}}}

\item{obs.cutoff}{\emph{Optional} Minimum number of observations required per pairwise relationship to be included in correlation matrix (default = \code{10}).}

\item{transformation}{\emph{Optional} Numeric transformation to apply to data (default = none). \code{"log10"} accepted.}

\item{out.dr}{Required. Output directory for pairwise summary data.}
}
\description{
This function filters input data and returns a normalized abundance matrix.
}
\details{
The parameter \code{obs.cutoff} refers to the minimum number of observations
per pairwise relationship required for inclusion in the output matrix.
}
\examples{
prepare_corr_data(prepared.data, obs.cutoff, transformation, out.dr)
}
