% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_corr_matrix.R
\name{prepare_corr_matrix}
\alias{prepare_corr_matrix}
\title{Convert preprocessed abundance data into pairwise correlation matrix}
\usage{
prepare_corr_matrix(prepared.data, transformation, correlation)
}
\arguments{
\item{prepared.data}{Output from \code{\link{prepare_network_data}} or \code{\link{split_network_data}}}

\item{transformation}{\emph{Optional} Numeric transformation to apply to relative abundance data (default = none). \code{"log10"} accepted.}

\item{correlation}{Metric for pairwise comparisons (see \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}})}
}
\value{

}
\description{
This function calculates pairwise correlation coefficients and associated p-values from pre-processed abundance data. This is essentially a convenience wrapper around \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}}.
}
