% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_net_features.R
\name{prepare_net_features}
\alias{prepare_net_features}
\title{Preprocess network edge data}
\usage{
prepare_net_features(
  edges.,
  node.ids,
  p.val.cutoff,
  correlation,
  correlation.cutoff,
  sign,
  num.cores = NULL
)
}
\arguments{
\item{edges.}{Output from \code{\link[=generate_edges]{generate_edges()}}}

\item{node.ids}{Output from \code{\link[=get_node_ids]{get_node_ids()}}}

\item{p.val.cutoff}{Any pairwise comparison with a p-value greater than this value will be dropped}

\item{correlation}{Metric for pairwise comparisions (see \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}})}

\item{correlation.cutoff}{Values from \code{0} to \code{1} accepted. Only relationships with an absolute value of this or greater are kept.}

\item{sign}{Direction(s) of pairwise relationship to include in network. Values from \code{"positive"}, \code{"negative"}, or \code{"all"} accepted.}

\item{num.cores}{See \code{\link[=check_cores]{check_cores()}}}
}
\description{
This function preprocesses edge and node data, combining both outputs into a list object to be used in
downstream functions, such as \code{\link[=get_igraph]{get_igraph()}}.
}
