% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_network_data.R
\name{split_network_data}
\alias{split_network_data}
\title{Split prepared data for separate networks}
\usage{
split_network_data(prepared.data, subset.column)
}
\arguments{
\item{prepared.data}{Output from \code{\link[=prepare_network_data]{prepare_network_data()}}.}

\item{subset.column}{Metadata variable by which \code{prepared.data} is split into separate dataframes for separate network analyses.}
}
\description{
This function produces a list of dataframes split by a categorical metadata variable
defined by \code{subset.column}. Comparisons within each \code{subset.column} group are dropped
if the number of observations is less than 5. Each split dataframe represents a
subcommunity from which separate networks will be constructed in downstream functions.
}
\examples{
split_network_data(prepared_data, "zone")
}
