% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cooccurrence.R
\name{summarize_cooccurrence}
\alias{summarize_cooccurrence}
\title{Summarize co-occurrences across subnetworks}
\usage{
summarize_cooccurrence(nodes., edges., tax.level)
}
\arguments{
\item{nodes.}{Node output from \code{bngal::extract_node_data}}

\item{edges.}{Network edge output data from \code{bngal::load_network_data}}

\item{tax.level}{Level of taxonomic classification}
}
\description{
Summarize co-occurrences across subnetworks
}
