% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ebc_abun.R
\name{calc_ebc_abun}
\alias{calc_ebc_abun}
\title{Calculate abundance of edge between clusters}
\usage{
calc_ebc_abun(ebc.nodes, binned.taxonomy, metadata.cols)
}
\value{
\strong{Deprecated!} This function will be removed in a future version of \code{bngal}. Please use \code{\link{ebc_compositions}} instead.
}
\description{
Calculate abundance of edge between clusters
}
