% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_betweenness.R
\name{get_edge_betweenness}
\alias{get_edge_betweenness}
\title{Return edge between cluster data}
\usage{
get_edge_betweenness(igraph.list)
}
\arguments{
\item{igraph.list}{Output from \code{\link[=get_igraph]{get_igraph()}}}
}
\value{

}
\description{
This is a wrapper around \code{\link[igraph:cluster_edge_betweenness]{igraph::cluster_edge_betweenness()}}. This can
be a memory- and time-intensive operation, especially for lower taxonomic
levels. In this wrapper, edge betweenness is processed in parallel for each
subcommunity by default in an attempt to ease runtime requirements. For larger
matrices of a single community, consider using \code{\link[parallel:mclapply]{parallel::mcmapply()}} around
\code{igraph::cluster_edge_betweenness()} to increase computational efficiency
(support coming soon).
}
\details{
Before running this function, consider
}
\examples{
get_igraph_output <- bngal::get_igraph(prepro_data)
get_edge_betweenness(get_igraph_output)
}
