% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_network_data.R
\name{load_network_data}
\alias{load_network_data}
\title{Load exported network data}
\usage{
load_network_data(network.dir, tax.level)
}
\arguments{
\item{network.dir}{Full path to network data output from either \verb{[bngal::export_network_data]}
(if running interactively) or bngal-build-networks.R (if using bngal CLI)}

\item{tax.level}{Taxonomic level of network data output from either \verb{[bngal::export_network_data]}
(if running interactively) or bngal-build-networks.R (if using bngal CLI)}
}
\description{
Load exported network data
}
