% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_comp.R
\name{plot_core_comp}
\alias{plot_core_comp}
\title{Plot edge between cluster compositions against node coreness}
\usage{
plot_core_comp(ebc.nodes.abun, tax.level, metadata, fill.by)
}
\arguments{
\item{ebc.nodes.abun}{Output from \code{\link[=ebc_compositions]{ebc_compositions()}}}

\item{tax.level}{Level of taxonomic classification at which to create dendrogram}

\item{metadata}{See \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{fill.by}{\emph{Optional} Metadata column by which to fill EBC composition plots}
}
\description{
Plot edge between cluster compositions against node coreness
}
