% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_network_data.R
\name{prepare_network_data}
\alias{prepare_network_data}
\title{Prepare ASV count data for correlation network analysis}
\usage{
prepare_network_data(binned.tax, meta.data, corr.cols, sub.comms)
}
\arguments{
\item{binned.tax}{Output from \code{\link[=bin_taxonomy]{bin_taxonomy()}}. Must be an absolute abundance ASV table.}

\item{meta.data}{See \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{corr.cols}{Metadata variables (such as sample environmental data) to include in network correlations. Variables must be supplied as a concatenated string of column names from \code{meta.data} (e.g., \code{c('var1','var2')}). Default = \code{NULL}}

\item{sub.comms}{\emph{Optional} Metadata column by which to separate networks.}
}
\value{

}
\description{
This function prepares data for correlation network analysis at a specified
level of taxonomic classification. Singleton ASVs are first removed from a
provided binned data (from \code{\link[=bin_taxonomy]{bin_taxonomy()}}) that are filtered either above or below
a given abundance cutoff value based on user input. Corresponding metadata columns
(such as a geochemical parameter for each sample) may be included in the
preprocessed network data by specifying the \code{corr.cols} option.
Microbial abundance and environmental metadata (if applicable) are normalized
on a scale of 0-1 for downstream correlation matrix construction.
}
\examples{

}
