% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_summary.R
\name{pw_summary}
\alias{pw_summary}
\title{Summarize and export pairwise data for each sample after quality control}
\usage{
pw_summary(
  corr.data,
  preprocessed.features,
  tax.level,
  out.dr,
  num.cores = NULL
)
}
\arguments{
\item{corr.data}{Output from \code{\link[=prepare_corr_data]{prepare_corr_data()}}}

\item{preprocessed.features}{Required. Output from \code{\link[=prepare_net_features]{prepare_net_features()}}}

\item{tax.level}{Taxonomic level at which to summarize pairwise data.}

\item{out.dr}{Required. Output directory for pairwise summary data. Should be the same path as defined in \code{\link[=prepare_corr_data]{prepare_corr_data()}}}

\item{num.cores}{See \code{\link{check_cores}}}
}
\description{
Summarize and export pairwise data for each sample after quality control
}
