% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_summary.R
\name{pw_summary}
\alias{pw_summary}
\title{Summarize and export pairwise data for each sample after quality control}
\usage{
pw_summary(binned.taxonomy, preprocessed.features, out.dr, cores = 1)
}
\arguments{
\item{binned.taxonomy}{Output from \code{\link[=bin_taxonomy]{bin_taxonomy()}}}

\item{preprocessed.features}{Required. Output from \code{\link[=prepare_net_features]{prepare_net_features()}}}

\item{out.dr}{Required. Output directory for pairwise summary data. Should be the same path as defined in \code{\link[=prepare_corr_data]{prepare_corr_data()}}}

\item{cores}{\emph{(Optional)} Number of CPUs. Default = 1}
}
\value{

}
\description{
Summarize and export pairwise data for each sample after quality control
}
